+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* Breaking Changes
+  * Make `shippingMethod` property on `BTThreeDSecureRequest` an enum instead of a string
+
 ## 5.0.0-beta2 (2021-01-20)
 * Add SPM support for `BraintreeDataCollector` and `BraintreeThreeDSecure`
 * Add SPM libraries for `KountDataCollector` and `PPRiskMagnes` to workaround Xcode bug (addresses #576)